views cleanup
divide views into subdirectories to have a more clear division of views with different purposes.
