+++ tfc_web/dashboard/urls.py
@@ -1,11 +1,12 @@
 from django.conf.urls import url
-from dashboard import views
+from dashboard.views import dashboard
+from dashboard.views.widgets import weather, station_board
 
 
 urlpatterns = [
-    url(r'^design/', views.design, name='dashboard-design'),
-    url(r'^layout/(?P<layout_id>\d+)/$', views.layout, name='layout'),
-    url(r'^layout/(?P<layout_id>\d+)/config/$', views.layout_config, name='dashboard-layout-config'),
-    url(r'^weather$', views.weather, name='dashboard-weather'),
-    url(r'^station_board$', views.station_board, name='station-board')
+    url(r'^design/', dashboard.design, name='dashboard-design'),
+    url(r'^layout/(?P<layout_id>\d+)/$', dashboard.layout, name='layout'),
+    url(r'^layout/(?P<layout_id>\d+)/config/$', dashboard.layout_config, name='dashboard-layout-config'),
+    url(r'^weather$', weather.weather, name='dashboard-weather'),
+    url(r'^station_board$', station_board.station_board, name='station-board')
 ]
